+++ .github/workflows/unit-testing.yml
@@ -1,12 +1,12 @@
-# This workflow performs unit testing on testing utility packages
+# This workflow performs unit testing on utility packages
 name: Unit Testing
 
 on:
   pull_request:
     branches: [ main ]
     paths: 
-      - 'utils/**.go'
-  # manuall trigger
+      - 'test/utils'
+  # manual trigger
   workflow_dispatch:
 
 jobs:
@@ -20,4 +20,4 @@ jobs:
       with:
         go-version: 1.*
 
-    - run: go test -v ./utils
+    - run: go test -v ./utils/...
