+++ test/common-tests/version.go
@@ -1,13 +1,13 @@
 package common
 
 import (
-	"edgex-snap-testing/utils"
+	"edgex-snap-testing/test/utils"
 	"testing"
 
 	"github.com/stretchr/testify/require"
 )
 
-func TestVersion(t *testing.T, snapName string) {
+func Version(t *testing.T, snapName string) {
 	require.Regexp(t,
 		"^([0-9]+).([0-9]+).([0-9]+).*$",
 		utils.SnapVersion(t, snapName),
