+++ test/suites/device-mqtt/common_test.go
@@ -1,10 +1,10 @@
 package test
 
 import (
-	"edgex-snap-testing/test/common"
+	"edgex-snap-testing/test/common-tests"
 	"testing"
 )
 
 func TestVersion(t *testing.T) {
-	common.TestVersion(t, thisSnap)
+	common.Version(t, thisSnap)
 }
